+++ demos/quickstart/rollup/rollup.config.js
@@ -5,8 +5,14 @@ export default {
     output: {
         file: "dist/index.js",
         format: "umd",
-        name: "quickstart"
+        name: "quickstart",
+        globals: {
+            "@hpcc-js/dgrid-shim": "@hpcc-js/dgrid-shim"
+        }
     },
+    external: [
+        "@hpcc-js/dgrid-shim"
+    ],
     plugins: [
         nodeResolve()
     ]