+++ packages/chart/src/XYAxis.ts
@@ -28,7 +28,7 @@ export class XYAxis extends SVGWidget {
         this.domainAxis = new Axis()
             .classed({ domain: true })
             .orientation_default("bottom")
-            .type_default("ordinal")
+            .type("ordinal")
             .overlapMode_default("stagger")
             .shrinkToFit_default("high")
             .extend_default(0)
@@ -36,7 +36,7 @@ export class XYAxis extends SVGWidget {
         this.valueAxis = new Axis()
             .classed({ value: true })
             .orientation_default("left")
-            .type_default("linear")
+            .type("linear")
             .shrinkToFit_default("high")
             ;
 
@@ -228,9 +228,11 @@ export class XYAxis extends SVGWidget {
         return this.valueAxis.scalePos(d);
     }
 
-    setScaleRange(width, height) {
-        this.xAxis.width(width);
-        this.yAxis.height(height);
+    getAxisSize(host: XYAxis): { width: number, height: number } {
+        return {
+            width: !this.xAxis ? host.xAxis.width() : this.xAxis.width(),
+            height: !this.yAxis ? host.yAxis.width() : this.yAxis.height()
+        };
     }
 
     calcMargin(_domNode, element, isHorizontal) {