+++ tests/test-comms/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.6"></a>
+## [2.0.6](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-comms@2.0.5...@hpcc-js/test-comms@2.0.6) (2018-10-30)
+
+**Note:** Version bump only for package @hpcc-js/test-comms
+
+
+
+
+
 <a name="2.0.5"></a>
 ## [2.0.5](https://github.com/GordonSmith/Visualization/compare/@hpcc-js/test-comms@2.0.4...@hpcc-js/test-comms@2.0.5) (2018-10-15)